requirements.txt
opencv-python==4.10.0.84
numpy==1.26.3
onnx==1.17.0
onnxruntime-gpu==1.20.0
torch==2.5.1
torchvision==0.20.1
flask==3.0.3
pyzmq==26.2.0